Add key props to index.jsx
Note: I think opening dropdown menus got jankier from my previous commit. If the dropdown menu is near the bottom of the page, so it has to show upward, there might be a fraction of a second where a second vertical scrollbar on the right appears (possibly because for a sec it renders below the dropdown button and causes a second scrollbar), and also the footer inches upward for some weird reason.